requirements.txt
Flask
Flask-Cors
numpy
tensorflow
Pillow
pymongo